requirements.txt
flask==2.0.1
requests==2.26.0
networkx==2.6.3
python-dotenv==0.19.0
Flask-CORS==4.0.0
gunicorn==21.2.0
